requirements.txt
aiohttp
aiohttp-security
aiohttp-session[secure]
aiosqlite
aiohttp-jinja2
PyYAML
tortoise-orm[asyncpg,aiomysql]